Rename ResponseCache => CacheStrategy, allow different strategies
