+++ src/ReplayServiceProvider.php
@@ -3,6 +3,7 @@
 namespace Kayrunm\Replay;
 
 use Illuminate\Support\ServiceProvider;
+use Kayrunm\Replay\Cache\CacheStrategy;
 use Kayrunm\Replay\Strategies\Strategy;
 
 class ReplayServiceProvider extends ServiceProvider
@@ -19,5 +20,6 @@ class ReplayServiceProvider extends ServiceProvider
         $this->mergeConfigFrom(__DIR__ . '/../config/replay.php', 'replay');
 
         $this->app->bind(Strategy::class, config('replay.strategy'));
+        $this->app->bind(CacheStrategy::class, config('replay.caching_strategy'));
     }
 }
