+++ tests/Unit/ReplayTest.php
@@ -5,10 +5,10 @@ namespace Tests\Unit;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Kayrunm\Replay\Cache\CacheStrategy;
 use Kayrunm\Replay\Exceptions\MatchingRequestStillExecuting;
 use Kayrunm\Replay\Replay;
 use Kayrunm\Replay\ReplayResponse;
-use Kayrunm\Replay\ResponseCache;
 use Kayrunm\Replay\Strategies\Strategy;
 use Mockery\MockInterface;
 use Tests\TestCase;
@@ -24,7 +24,7 @@ class ReplayTest extends TestCase
         parent::setUp();
 
         $this->strategy = $this->mock(Strategy::class);
-        $this->cache = $this->mock(ResponseCache::class);
+        $this->cache = $this->mock(CacheStrategy::class);
 
         $this->middleware = $this->app->make(Replay::class);
     }
